+++ js/main.js
@@ -1,5 +1,4 @@
-import {CANVAS_SIZE} from './util.js';
-import { generatePost } from './generate-post.js';
+import {CANVAS_SIZE} from './data.js';
+import { generatePost } from './util.js';
 
-const similarPost = Array.from({length: CANVAS_SIZE}, generatePost);
-console.log(similarPost);
+
